requirements.txt
asgiref==3.5.0
cffi==1.15.0
coverage==5.2.1
cryptography==36.0.2
Django==4.0.1
django-admin-interface==0.18.6
django-appconf==1.0.5
django-autocomplete-light==3.9.1
django-colorfield==0.6.3
django-countries==7.2.1
django-credit-cards==0.4.1
django-crispy-forms==1.14.0
django-datetime-widget2==0.9.5
django-encrypted-model-fields==0.6.5
django-extensions==3.1.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-utils-six==2.0
djangorestframework==3.13.1
install==1.3.5
Pillow==9.0.1
pycparser==2.21
pyparsing==3.0.7
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
django-jet-reboot
pandas
openpyxl
